add hook for data fetching
